Stellar mobile cleanup: wallet list and platform popups (#14578)
* wallet list

* cleanup

* have wallet popup help more with platform split

* EnterNamePopup -> EnterName

* EnterKeyPopup -> EnterKey

* overflow hidden and storyshots
