+++ shared/wallets/common/index.js
@@ -1,7 +1,7 @@
 // @flow
 export {default as AccountEntry} from './account-entry'
 export {default as AccountPageHeader} from './account-page-header'
-export {default as EnterNamePopup} from './enter-name-popup'
+export {default as EnterName} from './enter-name'
 export {default as MarkdownMemo} from './markdown-memo'
 export {default as ParticipantsRow} from './participants-row'
 export {default as SmallAccountID} from './small-account-id'
