+++ shared/wallets/export-secret-key/index.stories.js
@@ -8,11 +8,11 @@ const load = () => {
   Sb.storiesOf('Wallets', module).add('Export secret key', () => (
     <ExportSecretKey
       accountID={stringToAccountID('GBCCH4KHE5MUXXYSFCKJ3BRN4U3MTXOXD2GBJH5V7QF6OJ6S5R23DWYF')}
+      accountName="Money for Sneakers"
       onClose={Sb.action('onClose')}
       onLoadSecretKey={Sb.action('onLoadSecretKey')}
       secretKey="SETECASTRONOMY"
       username="cecileb"
-      walletName="Money for Sneakers"
     />
   ))
 }
